feat(KUI-1149): moved main couse info to seperate component, added special case for courses without course rounds
